@@ -1,19 +1,16 @@
 {
   "name": "react-router-bootstrap",
   "version": "0.18.1",
-  "description": "react-router and react-bootstrap compatible components",
+  "description": "Integration between React Router and React-Bootstrap",
   "main": "./lib/index.js",
   "scripts": {
     "prepublish": "npm run build",
-    "build": "babel src --out-dir=lib && webpack && COMPRESS=1 webpack && npm run bower-prepare",
+    "build": "babel src --out-dir=lib && webpack && webpack -p",
     "test": "npm run lint && karma start --single-run",
     "tdd": "karma start",
-    "visual-test": "open http://localhost:8080/public/visual#/ && webpack-dev-server --config webpack.test.config.babel.js",
-    "lint": "eslint ./",
-    "bower-prepare": "babel-node scripts/bower-prepare.js",
-    "patch": "release patch",
-    "minor": "release minor",
-    "major": "release major"
+    "visual-test": "open http://localhost:8080/ && webpack-dev-server --config webpack.visual.config.babel.js",
+    "lint": "eslint *.babel.js src tests",
+    "release": "release"
   },
   "repository": {
     "type": "git",
@@ -34,48 +31,53 @@
   },
   "homepage": "https://github.com/react-bootstrap/react-router-bootstrap",
   "peerDependencies": {
-    "react-bootstrap": ">=0.22.4",
-    "react-router": ">=0.13.1"
+    "react": ">=0.13.0 || >=0.14.0-rc1",
+    "react-router": ">=1.0.0-rc1"
   },
   "devDependencies": {
-    "babel": "^5.5.6",
-    "babel-core": "^5.5.6",
-    "babel-eslint": "^4.0.5",
-    "babel-loader": "^5.1.4",
-    "bootstrap": "^3.3.1",
-    "chai": "^3.0.0",
-    "colors": "^1.1.2",
-    "css-loader": "^0.15.3",
-    "eslint": "^1.0.0",
-    "eslint-config-airbnb": "0.0.7",
-    "eslint-plugin-babel": "^1.0.0",
-    "eslint-plugin-mocha": "^0.4.0",
-    "eslint-plugin-react": "^3.1.0",
-    "karma": "^0.13.3",
+    "babel": "^5.8.23",
+    "babel-core": "^5.8.24",
+    "babel-eslint": "^4.1.2",
+    "babel-loader": "^5.3.2",
+    "bootstrap": "^3.3.5",
+    "css-loader": "^0.18.0",
+    "es5-shim": "^4.1.13",
+    "eslint": "1.3.x",
+    "eslint-config-airbnb": "0.0.8",
+    "eslint-plugin-babel": "^2.1.1",
+    "eslint-plugin-mocha": "^0.5.1",
+    "eslint-plugin-react": "^3.3.2",
+    "file-loader": "^0.8.4",
+    "history": "^1.9.1",
+    "html-webpack-plugin": "^1.6.1",
+    "karma": "^0.13.9",
     "karma-cli": "^0.1.0",
     "karma-mocha": "^0.2.0",
-    "karma-mocha-reporter": "^1.0.2",
-    "karma-phantomjs-launcher": "^0.2.0",
-    "karma-webpack": "^1.5.0",
+    "karma-mocha-reporter": "^1.1.1",
+    "karma-phantomjs-launcher": "^0.2.1",
+    "karma-sinon-chai": "^1.1.0",
+    "karma-sourcemap-loader": "^0.3.5",
+    "karma-webpack": "^1.7.0",
     "less": "^2.5.1",
     "less-loader": "^2.2.0",
-    "lodash": "^3.10.0",
-    "mocha": "^2.1.0",
-    "mt-changelog": "^0.6.1",
+    "mocha": "^2.3.2",
+    "mt-changelog": "^0.6.2",
     "node-libs-browser": "^0.5.2",
-    "phantomjs": "^1.9.13",
-    "react": ">0.10.0",
-    "react-bootstrap": ">=0.22.4",
-    "react-router": ">=0.13.1",
-    "release-script": "^0.2.1",
-    "shelljs": "^0.5.1",
+    "phantomjs": "^1.9.18",
+    "react": "^0.14.0-rc1",
+    "react-bootstrap": "^0.25.100-react-pre.1",
+    "react-dom": "^0.14.0-rc1",
+    "react-router": "^1.0.0-rc1",
+    "release-script": "^0.2.7",
     "style-loader": "^0.12.3",
     "url-loader": "^0.5.6",
-    "webpack": "^1.4.15",
-    "webpack-dev-server": "^1.7.0",
-    "yargs": "^3.15.0"
+    "webpack": "^1.12.1",
+    "webpack-dev-server": "^1.10.1",
+    "yargs": "^3.25.0"
   },
   "files": [
+    "README",
+    "CHANGELOG.md",
     "lib"
   ],
   "release-script": {
