@@ -1,19 +1,4 @@
-import './phantom-shims';
-import 'mocha';
+import 'es5-shim';
 
-const chai = require('chai');
-chai.should();
-
-global.expect = chai.expect;
-global.assert = chai.assert;
-
-global.TestUtils = require('react/addons').addons.TestUtils;
-
-import './ButtonLink.spec.js';
-import './ListGroupItemLink.spec.js';
-import './MenuItemLink.spec.js';
-import './NavItemLink.spec.js';
-import './PageItemLink.spec.js';
-import './RouterOverlayTrigger.spec.js';
-import './ThumbnailLink.spec.js';
-import './bower-imports-module.spec.js';
+const testsContext = require.context('.', true, /\.spec\.js$/);
+testsContext.keys().forEach(testsContext);
