@@ -1,33 +1,33 @@
 import React from 'react';
-import {Link} from 'react-router';
 import Nav from 'react-bootstrap/lib/Nav';
 import NavItem from 'react-bootstrap/lib/NavItem';
-import NavItemLink from '../../src/NavItemLink';
+import {Link} from 'react-router';
 
-const NavItemVisual = React.createClass({
-  handleSelect(selectedKey) {
-    window.alert('selected ' + selectedKey);
-  },
-  render() {
-    return (
-      <div>
-        <Link to='home'>&lt;-- Back to Index</Link>
-        <h2>NavItemLink</h2>
-        <h3>Baseline (Raw React-Bootstrap)</h3>
-        <Nav bsStyle="pills" activeKey={1} onSelect={this.handleSelect}>
-          <NavItem eventKey={1} href="/home">NavItem 1 content</NavItem>
-          <NavItem eventKey={2} title="Item">NavItem 2 content</NavItem>
-          <NavItem eventKey={3} disabled={true}>NavItem 3 content</NavItem>
-        </Nav>
-        <h3>NavItemLink</h3>
-        <Nav bsStyle="pills">
-          <NavItemLink to='nav-item'>NavItemLink 1 content</NavItemLink>
-          <NavItemLink to='home'>NavItemLink 2 content</NavItemLink>
-          <NavItemLink to='home' disabled={true}>NavItemLink 3 content</NavItemLink>
-        </Nav>
-      </div>
-    );
-  }
-});
+import LinkContainer from '../../src/LinkContainer';
 
-export default NavItemVisual;
+export default () => (
+  <div>
+    <Link to="/home">Back to Index</Link>
+    <h2>NavItem</h2>
+
+    <h3>Baseline</h3>
+    <Nav bsStyle="pills" activeKey={1}>
+      <NavItem eventKey={1}>NavItem 1 content</NavItem>
+      <NavItem eventKey={2}>NavItem 2 content</NavItem>
+      <NavItem eventKey={3} disabled>NavItem 3 content</NavItem>
+    </Nav>
+
+    <h3>LinkContainer</h3>
+    <Nav bsStyle="pills">
+      <LinkContainer to="/nav-item">
+        <NavItem>NavItem 1 content</NavItem>
+      </LinkContainer>
+      <LinkContainer to="/home">
+        <NavItem>NavItem 2 content</NavItem>
+      </LinkContainer>
+      <LinkContainer to="/home" disabled>
+        <NavItem>NavItem 3 content</NavItem>
+      </LinkContainer>
+    </Nav>
+  </div>
+);
